requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
SQLAlchemy==1.4.23
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-JWT-Extended==4.3.1
Flask-Bcrypt==0.7.1
Flask-WTF==1.0.0
requests==2.26.0
python-dotenv==0.19.1
email-validator==1.1.3
Flask-Limiter==2.4.0
gunicorn==20.1.0
pytest==6.2.5